fix: afterViewInit subscription
- `first()` breaks our current implementations of a custom DateTime
input which we receive a default text value and the DateTime input will
try to convert it to a proper Date. `first()` skips the second emission
of `valueChanges` when the proper Date is constructed
- `filter()` to filter out `empty string`, `null`, and `undefined` since
this subscription runs in afterViewInit (ideally once) so if the value
is `''/undefined/null`, it is ok to skip with `filter()`. The subsequent
changes are captured by `HostListener`
- `take(1)` to complete the stream which effectively unsubscribes the
stream after one emission.